Fixed Tiled Map + Player Confine Movement
The cubes now take the positions they should, I was calculating the position of a cube and assigning it to the next one, so the missing cube wasn't missing, it was assigned the position of the previous one.
I also modified the ConfineMovement function on the PlayerMovement.cs so it works correctly.