Tribes Part 1
The tribe now has the same setup as the player minus the player input and sending data for display.
The functionality of buttons that was previously only supported for the player (had to be calling it with player tribe index) now is available for the rest of the tribes independent of index.